Multi-Level-RL-Trading-Systems/ppo_baseline/requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gym>=0.26.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.0.0

# Technical Analysis
ta-lib>=0.4.0
yfinance>=0.2.0

# Data Processing
ccxt>=4.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Testing
pytest>=6.0.0
pytest-cov>=3.0.0

# Development
black>=22.0.0
flake8>=4.0.0
isort>=5.0.0
